Suspend running CI on Ruby 2.0 until backports bug is fixed.
Currently there is a bug in Ruby 1.9.1 backports IO lib that causes a bad file descriptor error.
Sndicate is developed against Ruby 2.0.0 on multiple machines and is known
to work on it. It's unfortunate that the test suite can't run no the
platform it was developed on right now.

Will return to this issue before the first RC is released.