Add configurable posts route.
Posts will be behind a /prefix route to avoid querying the database for
both pages and posts when hitting and /* route. The prefix is configurable
in the main config.yml file. This change lets users hit the posts controller
via the user's chosen prefix as well as /posts/posts_path.

In the future, we'll allow the prefix path to be empty which removes the
prefix (which some consider ugly) from post paths but will add server load.